+++ test/unit/learning_log_scale_fixture_test.dart
@@ -1,5 +1,6 @@
 import 'package:ayutam/core/id/id_generator.dart';
 import 'package:ayutam/core/time/clock_service.dart';
+import 'package:ayutam/core/time/timezone_service.dart';
 import 'package:ayutam/database/app_database.dart';
 import 'package:ayutam/features/learning_log/application/learning_log_service.dart';
 import 'package:ayutam/features/learning_log/application/session_note_service.dart';
@@ -48,6 +49,7 @@ void main() {
         indexer: indexer,
         uow: uow,
         clock: clock,
+        timezones: const FakeTimezoneService(),
         ids: ids,
         deviceId: db.requireDeviceId,
       );